+++ arch/i386/cpu.h
@@ -867,17 +867,17 @@ int cpu_x86_handle_mmu_fault(CPUState *env, target_ulong addr,
 #define cpu_handle_mmu_fault cpu_x86_handle_mmu_fault
 void cpu_x86_set_a20(CPUState *env, int a20_state);
 
-static inline int hw_breakpoint_enabled(unsigned long dr7, int index)
+static inline int hw_breakpoint_enabled(uintptr_t dr7, int index)
 {
     return (dr7 >> (index * 2)) & 3;
 }
 
-static inline int hw_breakpoint_type(unsigned long dr7, int index)
+static inline int hw_breakpoint_type(uintptr_t dr7, int index)
 {
     return (dr7 >> (DR7_TYPE_SHIFT + (index * 4))) & 3;
 }
 
-static inline int hw_breakpoint_len(unsigned long dr7, int index)
+static inline int hw_breakpoint_len(uintptr_t dr7, int index)
 {
     int len = ((dr7 >> (DR7_LEN_SHIFT + (index * 4))) & 3);
     return (len == 2) ? 8 : len + 1;
