+++ include/cpu-common.h
@@ -12,7 +12,7 @@
 #include "tlib-queue.h"
 
 /* address in the RAM (different from a physical address) */
-typedef unsigned long ram_addr_t;
+typedef uintptr_t ram_addr_t;
 
 /* memory API */
