+++ tcg/tcg.c
@@ -102,7 +102,7 @@ static inline void tcg_out32(TCGContext *s, uint32_t v)
 /* label relocation processing */
 
 static void tcg_out_reloc(TCGContext *s, uint8_t *code_ptr, int type,
-                          int label_index, long addend)
+                          int label_index, uintptr_t addend)
 {
     TCGLabel *l;
     TCGRelocation *r;
@@ -291,8 +291,8 @@ void tcg_prologue_init()
     tcg->ctx->code_buf = tcg->code_gen_prologue;
     tcg->ctx->code_ptr = tcg->ctx->code_buf;
     tcg_target_qemu_prologue(tcg->ctx);
-    flush_icache_range((unsigned long)tcg->ctx->code_buf,
-                       (unsigned long)tcg->ctx->code_ptr);
+    flush_icache_range((uintptr_t)tcg->ctx->code_buf,
+                       (uintptr_t)tcg->ctx->code_ptr);
 }
 
 void tcg_set_frame(TCGContext *s, int reg,
@@ -1751,8 +1751,7 @@ static int tcg_reg_alloc_call(TCGContext *s, const TCGOpDef *def,
     return nb_iargs + nb_oargs + def->nb_cargs + 1;
 }
 
-static inline int tcg_gen_code_common(TCGContext *s, uint8_t *gen_code_buf,
-                                      long search_pc)
+static inline int tcg_gen_code_common(TCGContext *s, uint8_t *gen_code_buf, uintptr_t search_pc)
 {
     TCGOpcode opc;
     int op_index;
@@ -1813,7 +1812,7 @@ static inline int tcg_gen_code_common(TCGContext *s, uint8_t *gen_code_buf,
             break;
         case INDEX_op_set_label:
             tcg_reg_alloc_bb_end(s, s->reserved_regs);
-            tcg_out_label(s, args[0], (long)s->code_ptr);
+            tcg_out_label(s, args[0], (uintptr_t)s->code_ptr);
             break;
         case INDEX_op_call:
             dead_args = s->op_dead_args[op_index];
@@ -1849,8 +1848,8 @@ int tcg_gen_code(TCGContext *s, uint8_t *gen_code_buf)
     tcg_gen_code_common(s, gen_code_buf, -1);
 
     /* flush instruction cache */
-    flush_icache_range((unsigned long)gen_code_buf, 
-                       (unsigned long)s->code_ptr);
+    flush_icache_range((uintptr_t)gen_code_buf, 
+                       (uintptr_t)s->code_ptr);
     return s->code_ptr -  gen_code_buf;
 }
 
@@ -1858,7 +1857,7 @@ int tcg_gen_code(TCGContext *s, uint8_t *gen_code_buf)
    offset bytes from the start of the TB.  The contents of gen_code_buf must
    not be changed, though writing the same values is ok.
    Return -1 if not found. */
-int tcg_gen_code_search_pc(TCGContext *s, uint8_t *gen_code_buf, long offset)
+int tcg_gen_code_search_pc(TCGContext *s, uint8_t *gen_code_buf, uintptr_t offset)
 {
     return tcg_gen_code_common(s, gen_code_buf, offset);
 }
